barebones_starter/requirements.txt
# --- Core LLM & LangChain stack ---
openai>=1.51.0
langchain>=1.0.0,<2.0.0
langchain-openai==1.0.2

# --- Web UI & Environment ---
streamlit>=1.38.0
python-dotenv>=1.0.1
pandas>=2.2.2

# --- Optional but recommended ---
tiktoken>=0.8.0
requests>=2.32.3
tenacity>=8.2.3
